Move browser process execution functions to BrowserProcess
These are not specific to Chromium, and are better suited to be used by
BrowserProcess.

parseDevToolsURL is arguably Chrome-specific, but support for other
browsers could be added, and it could hopefully be rewritten to not
involve parsing stdout...
